requirements.txt
aio-pika==9.3.0
aiofiles==23.1.0
aiogram==2.19
aiohttp==3.8.6
aiormq==6.7.7
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
idna==3.4
jsonpointer==2.0
magic-filter==1.0.12
msgpack==1.0.7
multidict==6.0.4
pamqp==3.2.1
Pillow==9.1.0
pydantic==2.3.0
pydantic_core==2.6.3
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
typing_extensions==4.7.1
yarl==1.9.2
